+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   ([Issue #671](https://github.com/scoutapp/scout_apm_python/issues/671))
 
 ### Fixed
+- Switch to storing TrackedRequest in a ContextVar for Python3.6+ rather than in a thread local.
 
 ## [2.21.1] 2021-07-27